+++ src/AIPrompt.App/Services/IAutoBackupService.cs
@@ -1,0 +1,8 @@
+namespace AIPrompt.App.Services;
+
+public interface IAutoBackupService
+{
+    void Start();
+
+    void Reconfigure();
+}