+++ src/AIPrompt.Core/Models/ImportMode.cs
@@ -1,0 +1,7 @@
+namespace AIPrompt.Core.Models;
+
+public enum ImportMode
+{
+    Merge,
+    Overwrite
+}